alquinas/pyproject.toml
[tool.poetry]
name = "alquinas_api"
version = "0.1.0"
description = ""
authors = ["Daniel Cabrera <danielmtv6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.68"
uvicorn = "^0.15"
alembic = "^1.6"
sqlalchemy = "^1.4"
mysql-connector-python = "^8.0"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
